+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.x.x - 2021-04-27
 
 ### Added
+- Added helper filter-object
+
+### Changed
+- Changed QasProfile layout and structure
 - Added modifier class `qas-table--mobile`
 ## 1.x.x - 2021-04-29